@@ -6,5 +6,5 @@ export 'model.dart';
 
 class Geocoder {
   static final Geocoding local = LocalGeocoding();
-  static Geocoding google(String apiKey, { String language }) => GoogleGeocoding(apiKey, language: language);
+  static Geocoding google(String apiKey, { String? language }) => GoogleGeocoding(apiKey, language: language);
 }
